Fix build on Debian
In clean Debian unstable build chroots, the udisks plugin
of mate-sensors-applets does not get built.

Reason: The pkg-config tool returns an empty string for
"pkg-config --cflags libatasmart" in clean Debian unstable
build chroots. This probably is the case, because the result
"-I/usr/include" is in the include path anyway and would not
add include locations to the build process.

Checking the LIBATASMART_LIBS variable is more safe here
(as opposed to checking LIBATASMART_CFLAGS). It always contains
a value if the library is installed and has a proper version.

Closes https://github.com/mate-desktop/mate-sensors-applet/issues/17